@@ -2,8 +2,10 @@ group 'com.chendexiaqu'
 version '1.0-SNAPSHOT'
 
 apply plugin: 'java'
+apply plugin: 'idea'
 
-sourceCompatibility = 1.5
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 repositories {
     mavenCentral()
@@ -11,4 +13,6 @@ repositories {
 
 dependencies {
     testCompile group: 'junit', name: 'junit', version: '4.11'
+    compile 'com.google.guava:guava:11.0.2'
+    compile 'org.apache.httpcomponents:httpclient:4.5.2'
 }
